Declare type of int for Eyear

+++ src/OrbitalParameters.cpp
@@ -72,7 +72,7 @@ OrbitalParameters::OrbitalParameters() {
 
 void
 OrbitalParameters::calculateTime(const TLE &tle, char typerun, char typeinput, char opsmode, gravconsttype whichconst) {
-    int Eyear;
+    int_fast16_t Eyear;
     int timeDay;
     int mon;
     int day;
